+++ sbin/start_validator.sh
@@ -6,6 +6,11 @@ SBIN="`cd "$SBIN"; pwd`"
 cd $DATA_DIR
 
 args=(
+    #--datadir ./data_sequencer
+    #--http --http.addr '0.0.0.0' --http.port 4011 --http.api 'personal,eth,net,web3,txpool,miner,proof,debug'
+    #--ws --ws.addr '0.0.0.0' --ws.port 4012 --ws.api 'personal,eth,net,web3,txpool,miner,proof,debug'
+    #--http.corsdomain '*' --ws.origins '*'
+    #--networkid $NETWORK_ID
     --rollup.sequencer
     --rollup.l1.endpoint $L1_ENDPOINT
     --rollup.l1.chainid $L1_CHAIN_ID
