--- services/sidecar/data/.gitignore
@@ -1,7 +1,0 @@
-data
-data_sequencer
-data_validator
-data_indexer
-data_clef
-genesis.json
-initialVmHash.txt
