--- services/sidecar/data/keys/sequencer.prv
@@ -1,1 +1,0 @@
-ac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80
