--- services/sidecar/data/keys/validator.prv
@@ -1,1 +1,0 @@
-59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d
