--- services/sidecar/data/password.txt
@@ -1,1 +1,0 @@
-
