--- services/sidecar/internal/debug/trace.go
@@ -1,64 +1,0 @@
-// Copyright 2016 The go-ethereum Authors
-// This file is part of the go-ethereum library.
-//
-// The go-ethereum library is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Lesser General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// The go-ethereum library is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU Lesser General Public License for more details.
-//
-// You should have received a copy of the GNU Lesser General Public License
-// along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
-
-//go:build go1.5
-// +build go1.5
-
-package debug
-
-import (
-	"errors"
-	"os"
-	"runtime/trace"
-
-	"github.com/ethereum/go-ethereum/log"
-)
-
-// StartGoTrace turns on tracing, writing to the given file.
-func (h *HandlerT) StartGoTrace(file string) error {
-	h.mu.Lock()
-	defer h.mu.Unlock()
-	if h.traceW != nil {
-		return errors.New("trace already in progress")
-	}
-	f, err := os.Create(expandHome(file))
-	if err != nil {
-		return err
-	}
-	if err := trace.Start(f); err != nil {
-		f.Close()
-		return err
-	}
-	h.traceW = f
-	h.traceFile = file
-	log.Info("Go tracing started", "dump", h.traceFile)
-	return nil
-}
-
-// StopTrace stops an ongoing trace.
-func (h *HandlerT) StopGoTrace() error {
-	h.mu.Lock()
-	defer h.mu.Unlock()
-	trace.Stop()
-	if h.traceW == nil {
-		return errors.New("trace not in progress")
-	}
-	log.Info("Done writing Go trace", "dump", h.traceFile)
-	h.traceW.Close()
-	h.traceW = nil
-	h.traceFile = ""
-	return nil
-}
